linux-omap4: fix staging

+++ oe/recipes/linux/linux-omap4_3.4.23-1489.18.bb
@@ -56,10 +56,12 @@ do_install_append() {
 }
 
 staging_helper_append() {
-	rm ${STAGING_KERNEL_DIR}/include/linux/omap_drm.h
-	rm ${STAGING_KERNEL_DIR}/include/linux/omap_drv.h
-	cp -L ${S}/include/linux/omap_drm.h ${STAGING_KERNEL_DIR}/include/linux
-	cp -L ${S}/include/linux/omap_drv.h ${STAGING_KERNEL_DIR}/include/linux
+	rm -f ${STAGING_KERNEL_DIR}/include/linux/omap_drm.h
+	rm -f ${STAGING_KERNEL_DIR}/include/linux/omap_drv.h
+	if [ -f cp -L ${S}/include/linux/omap_drm.h ]; then
+		cp -L ${S}/include/linux/omap_drm.h ${STAGING_KERNEL_DIR}/include/linux
+		cp -L ${S}/include/linux/omap_drv.h ${STAGING_KERNEL_DIR}/include/linux
+	fi
 }
 
 PACKAGES =+ "kernel-headers"
